enrollments edits, until not working

+++ src/Kanbas/Courses/Assignments/index.tsx
@@ -14,51 +14,72 @@ export default function Assignments() {
   const { cid } = useParams();
   const [assignmentName, setAssignmentName] = useState("");
   const assignments = useSelector((state: any) => state.assignmentsReducer.assignments);
- 
+
   const { modules } = useSelector((state: any) => state.assignmentsReducer);
   const dispatch = useDispatch();
+
+  const formatDate = (dateString: string) => {
+    const date = new Date(dateString);
+    return date.toLocaleDateString('en-US', {
+      year: 'numeric',
+      month: 'long',
+      day: 'numeric',
+      hour: 'numeric',
+      minute: '2-digit',
+      hour12: true
+    });
+  };
+
+  const handleDeleteAssignment = (assignmentId: string) => {
+    if (window.confirm("Are you sure you want to remove this assignment?")) {
+      dispatch(deleteAssignment(assignmentId));
+    }
+  };
+
+ 
   return (
     <div id="wd-assignments">
       <div className="row">
-      <AssignmentControl />
+        <AssignmentControl />
       </div>
       <ul className="list-group rounded-0">
-      <li className="wd-module list-group-item p-0
+        <li className="wd-module list-group-item p-0
                    mb-5 fs-5 border-gray dropdown">
           <div className="wd-title p-3 ps-2 bg-secondary">
-          <BsGripVertical className="me-2 fs-3" />
-          <div className="btn dropdown-toggle"><strong>ASSIGNMENTS</strong></div>
-          <AssignmentListControlButtons /> 
-          <div className="ellipse float-end">40% of Total </div>
+            <BsGripVertical className="me-2 fs-3" />
+            <div className="btn dropdown-toggle"><strong>ASSIGNMENTS</strong></div>
+            <AssignmentListControlButtons />
+            <div className="ellipse float-end">40% of Total </div>
           </div>
           <ul className="wd-assignments-list list-group rounded-0">
-      {assignments
+            {assignments
 
-          .filter((assignment: any) => assignment.course === cid)
-          .map((assignment: any) => (
-            <li className="wd-assignment-list-item list-group-item p-3 ps-1">
-            <div>
-                <Link to={`/Kanbas/Courses/${cid}/Assignments/${assignment._id}`}
-                  className="wd-assignment-link indented-text assignment-link"
-                   >
-                  {assignment.title}
-                </Link><br /><BsGripVertical className="me-2 fs-3" /><MdAssignment className="me-4 fs-3 text-success"/>
-                <Link to={`/Kanbas/Courses/${cid}/Assignments/${assignment._id}`} className="wd-assignment-link text-danger multiple-modules-link" >
-                Multiple Modules
-                </Link>  | <strong>Not available until</strong> {assignment.until} | <AssignmentControlButtons assignmentId={assignment._id}
-          deleteAssignment={(assignmentId) => {
-            dispatch(deleteAssignment(assignmentId));
-          }}/><br />
-                <div className="indented-text"><strong>Due</strong> {assignment.due_assign} | {assignment.points} pts</div>
-                
-              </div>
-              
-            </li>
-      
-        ))}
-        </ul>
+              .filter((assignment: any) => assignment.course === cid)
+              .map((assignment: any) => (
+                <li className="wd-assignment-list-item list-group-item p-3 ps-1">
+                  <div>
+                    <Link to={`/Kanbas/Courses/${cid}/Assignments/${assignment._id}`}
+                      className="wd-assignment-link indented-text assignment-link"
+                    >
+                      {assignment.title}
+                    </Link><br /><BsGripVertical className="me-2 fs-3" /><MdAssignment className="me-4 fs-3 text-success" />
+                    <Link to={`/Kanbas/Courses/${cid}/Assignments/${assignment._id}`} className="wd-assignment-link text-danger multiple-modules-link" >
+                      Multiple Modules
+                    </Link>  | <strong>Not available until</strong> {formatDate(assignment.availability)} | <AssignmentControlButtons assignmentId={assignment._id}
+                      //deleteAssignment={(assignmentId) => {
+                        //dispatch(deleteAssignment(assignmentId));}}
+                        deleteAssignment={() => handleDeleteAssignment(assignment._id)}
+                       /><br />
+                    <div className="indented-text"><strong>Due</strong> {formatDate(assignment.due)} | {assignment.points} pts</div>
+
+                  </div>
+
+                </li>
+
+              ))}
+          </ul>
         </li>
-        
+
       </ul>
       {/*
       <ul className="list-group rounded-0">
@@ -116,7 +137,7 @@ export default function Assignments() {
       </ul>*/}
 
 
-      
+
     </div>
   );
 }
